Moved all sound logic into one file
It's a waste of resources to constantly add and destroy sound objects
every time we need to play a sound. These new changes now have the
sounds only initialize once per scene, and are stored in a sound_bank
which can be used to repeat the sounds.

A mute button (m) was also added because I was losing my mind listening
to that song

Moved player timing vars into the timers object in the game scene