Added IconButton class
Currently, we don't need an IconButton yet. However, when we add the
temporary (or permanent?) powerups to the game, such as spread shot,
nuke, penetrating shot, etc. These powerups will not have a numerical
value associated with Shermie and therefore, a spinner component would
not make sense to use. Later, when we get around to adding these
powerups, we'll need this.