Added "Mute Sounds: M" to `HowToPlay.js`
I'm guessing Jordan coded in the mute sounds thing when she was testing
because she didn't know that the M key actually mutes sounds. Either
way, this information should be relayed to the player when they view the
How To Play screen, so this was added.

Also formatted it a bit, again.