Muting enabled in menus and with esc/p
There is an issue with when you start a game, then leave the game and
unmute the sounds. The song isn't supposed to play there, but in reality
this will work out when we have a main menu song. It resumes the ingame
song because that's the only song we have in the game atm