Hotfix: Fixed Bug Where Powerup Bar Shows With no Powerups
Previously, you could reproduce this bug by doing the following:
    1. pick up powerup
    2. leave the game
    3. enter either play or sandbox scene

This happened because I previously implemented the variable that held
powerups as static, when it should not have had its lifetime be
persisent.

I'm pushing this to main because this change only required changing a
few lines to rename a variable.