Fixed Dialogue Freezing When Pausing
Previously, the dialogue would freeze if the player used esc to pause
the game while dialogue was being processed. This is due to the fact
that ESC was used for pausing and skipping dialogue, so pausing with ESC
would trigger both events, and cause the dialogue to be skipped while
paused, resulting in frozen dialogue that won't go away.

Unfortunately, even though I prefer to pause with `ESC`, we can no
longer have `ESC` for pausing unless we change the skip dialogue button
to something else. I am open to ideas.