Added Yes/No Dialog For Level Select
Yes/No Dialog asks the player if they wish to start the level with money
or not. Currently, I'm giving the player a little less than half of what
they would have gotten playing the game normally (not including rewards
from the USB enemy). This is subject to change later.